refactor: extract HeaderBar, LoadingOverlay, ErrorToast from App.svelte (705→222 lines)
