+++ src/App.svelte
@@ -1,16 +1,14 @@
 <script lang="ts">
-import QRCode from "qrcode";
 import { onMount } from "svelte";
-// biome-ignore lint/correctness/noUnusedImports: Used in Svelte template
+import ErrorToast from "./components/ui/ErrorToast.svelte";
 import FileUpload from "./components/ui/FileUpload.svelte";
-// biome-ignore lint/correctness/noUnusedImports: Used in Svelte template
+import HeaderBar from "./components/ui/HeaderBar.svelte";
+import LoadingOverlay from "./components/ui/LoadingOverlay.svelte";
 import ViewerCanvas from "./components/viewer/ViewerCanvas.svelte";
 import { DEBUG } from "./lib/config";
 import { loadGDSIIFromBuffer } from "./lib/utils/gdsLoader";
 import { fetchGDSIIFromURL } from "./lib/utils/urlLoader";
-// biome-ignore lint/correctness/noUnusedImports: Used in template via $collaborationStore
 import { collaborationStore } from "./stores/collaborationStore";
-// biome-ignore lint/correctness/noUnusedImports: Used in template via $gdsStore
 import { gdsStore } from "./stores/gdsStore";
 
 /**
@@ -131,152 +129,10 @@ onMount(async () => {
 		}
 	}
 });
-
-/**
- * Handle creating a collaboration session
- */
-function handleCreateSession() {
-	// Create session without requiring a file first
-	// File will be uploaded to session when user uploads it
-	collaborationStore.createSession();
-
-	if (DEBUG) {
-		console.log("[App] Session created. Upload a file to share it with peers.");
-	}
-}
-
-/**
- * Handle leaving a collaboration session
- */
-function handleLeaveSession() {
-	collaborationStore.leaveSession();
-}
-
-/**
- * Copy session link to clipboard
- */
-async function copySessionLink() {
-	const sessionId = $collaborationStore.sessionId;
-	if (!sessionId) return;
-
-	const url = new URL(window.location.href);
-	url.searchParams.set("room", sessionId);
-	const link = url.toString();
-
-	try {
-		await navigator.clipboard.writeText(link);
-		if (DEBUG) {
-			console.log("[App] Copied session link to clipboard:", link);
-		}
-		// TODO: Show success toast
-	} catch (error) {
-		console.error("[App] Failed to copy session link:", error);
-		gdsStore.setError("Failed to copy session link to clipboard");
-	}
-}
-
-/**
- * QR code state and functions
- */
-let showQRCode = false;
-let qrCodeDataUrl = "";
-
-/**
- * Toggle QR code panel
- */
-async function toggleQRCode() {
-	showQRCode = !showQRCode;
-
-	if (showQRCode && !qrCodeDataUrl) {
-		// Generate QR code
-		const sessionId = $collaborationStore.sessionId;
-		if (!sessionId) return;
-
-		const url = new URL(window.location.href);
-		url.searchParams.set("room", sessionId);
-		const link = url.toString();
-
-		try {
-			qrCodeDataUrl = await QRCode.toDataURL(link, {
-				width: 300,
-				margin: 2,
-				color: {
-					dark: "#000000",
-					light: "#FFFFFF",
-				},
-			});
-			if (DEBUG) {
-				console.log("[App] Generated QR code for session link");
-			}
-		} catch (error) {
-			console.error("[App] Failed to generate QR code:", error);
-			gdsStore.setError("Failed to generate QR code");
-		}
-	}
-}
-
-/**
- * Reset QR code when leaving session
- */
-$: if (!$collaborationStore.isInSession) {
-	showQRCode = false;
-	qrCodeDataUrl = "";
-}
 </script>
 
 <main class="app-main">
-	<div class="header">
-		<div class="header-content">
-			<div class="title-section">
-				<a href="/" class="title-container">
-					<img src="/icon.svg" alt="GDSJam" class="title-icon" />
-					<h1 class="title">GDSJam</h1>
-				</a>
-				<p class="subtitle">Collaborative GDSII Viewer</p>
-				{#if $gdsStore.fileName}
-					<p class="file-name">Loaded: {$gdsStore.fileName}</p>
-				{/if}
-			</div>
-
-			<div class="session-controls">
-				{#if $collaborationStore.isInSession}
-					<div class="session-info">
-						<span class="session-label">Session Active</span>
-						<span class="session-id">{$collaborationStore.sessionId?.substring(0, 8)}...</span>
-						<span class="user-count">{$collaborationStore.connectedUsers.length} user{$collaborationStore.connectedUsers.length !== 1 ? 's' : ''}</span>
-					</div>
-					<button class="btn btn-secondary" onclick={copySessionLink}>
-						Copy Link
-					</button>
-					<button class="btn btn-secondary" onclick={toggleQRCode}>
-						{showQRCode ? 'Hide QR Code' : 'Show QR Code'}
-					</button>
-					<button class="btn btn-danger" onclick={handleLeaveSession}>
-						Leave Session
-					</button>
-				{:else}
-					<button class="btn btn-primary" onclick={handleCreateSession}>
-						Create Session
-					</button>
-				{/if}
-			</div>
-		</div>
-	</div>
-
-	<!-- QR Code Panel - positioned as overlay, not in banner -->
-	{#if $collaborationStore.isInSession && showQRCode}
-		<div class="qr-code-panel">
-			<div class="qr-code-content">
-				<h3 class="qr-code-title">Scan to Join Session</h3>
-				{#if qrCodeDataUrl}
-					<img src={qrCodeDataUrl} alt="Session QR Code" class="qr-code-image" />
-					<p class="qr-code-hint">Scan this QR code with your mobile device to join the session</p>
-				{:else}
-					<p class="qr-code-loading">Generating QR code...</p>
-				{/if}
-			</div>
-		</div>
-	{/if}
+	<HeaderBar />
 
 	<div class="viewer-wrapper">
 		{#if !$gdsStore.document && !$gdsStore.isLoading}
@@ -284,24 +140,10 @@ $: if (!$collaborationStore.isInSession) {
 				<FileUpload />
 			</div>
 		{:else if $gdsStore.isLoading || $gdsStore.isRendering || $collaborationStore.isTransferring}
-			<div class="loading-overlay">
-				<div class="loading-content">
-					<div class="spinner"></div>
-					{#if $collaborationStore.isTransferring}
-						<p class="loading-message">{$collaborationStore.fileTransferMessage}</p>
-						<div class="progress-bar">
-							<div class="progress-fill" style="width: {$collaborationStore.fileTransferProgress}%"></div>
-						</div>
-						<p class="progress-text">{Math.round($collaborationStore.fileTransferProgress)}%</p>
-					{:else}
-						<p class="loading-message">{$gdsStore.loadingMessage}</p>
-						<div class="progress-bar">
-							<div class="progress-fill" style="width: {$gdsStore.loadingProgress}%"></div>
-						</div>
-						<p class="progress-text">{Math.round($gdsStore.loadingProgress)}%</p>
-					{/if}
-				</div>
-			</div>
+			<LoadingOverlay
+				message={$collaborationStore.isTransferring ? $collaborationStore.fileTransferMessage : $gdsStore.loadingMessage}
+				progress={$collaborationStore.isTransferring ? $collaborationStore.fileTransferProgress : $gdsStore.loadingProgress}
+			/>
 		{/if}
 
 		{#if $gdsStore.document}
@@ -309,15 +151,7 @@ $: if (!$collaborationStore.isInSession) {
 		{/if}
 
 		{#if $gdsStore.error}
-			<div class="error-overlay">
-				<div class="error-content">
-					<p class="error-title">Error</p>
-					<p class="error-message">{$gdsStore.error}</p>
-					<button class="error-button" onclick={() => gdsStore.clearError()}>
-						Dismiss
-					</button>
-				</div>
-			</div>
+			<ErrorToast message={$gdsStore.error} onDismiss={() => gdsStore.clearError()} />
 		{/if}
 	</div>
 
@@ -340,137 +174,6 @@ $: if (!$collaborationStore.isInSession) {
 		background-color: #1a1a1a;
 	}
 
-	.header {
-		padding: 1rem 1.5rem;
-		background-color: #0f0f0f;
-		border-bottom: 1px solid #333;
-	}
-
-	.header-content {
-		display: flex;
-		justify-content: space-between;
-		align-items: center;
-		gap: 2rem;
-	}
-
-	.title-section {
-		flex: 1;
-	}
-
-	.title-container {
-		display: flex;
-		align-items: center;
-		gap: 0.75rem;
-		text-decoration: none;
-	}
-
-	.title-container:hover .title {
-		color: #6bb3ff;
-	}
-
-	.title-icon {
-		width: 2rem;
-		height: 2rem;
-	}
-
-	.title {
-		margin: 0;
-		font-size: 1.5rem;
-		font-weight: 600;
-		color: #fff;
-	}
-
-	.subtitle {
-		margin: 0.25rem 0 0 0;
-		font-size: 0.875rem;
-		color: #888;
-	}
-
-	.file-name {
-		margin: 0.5rem 0 0 0;
-		font-size: 0.875rem;
-		color: #4a9eff;
-		font-weight: 500;
-	}
-
-	.session-controls {
-		display: flex;
-		align-items: center;
-		gap: 0.75rem;
-	}
-
-	.session-info {
-		display: flex;
-		align-items: center;
-		gap: 0.5rem;
-		padding: 0.5rem 1rem;
-		background-color: #1a1a1a;
-		border: 1px solid #333;
-		border-radius: 6px;
-	}
-
-	.session-label {
-		font-size: 0.75rem;
-		color: #888;
-		text-transform: uppercase;
-		letter-spacing: 0.05em;
-	}
-
-	.session-id {
-		font-size: 0.875rem;
-		color: #4a9eff;
-		font-family: monospace;
-	}
-
-	.user-count {
-		font-size: 0.875rem;
-		color: #4ecdc4;
-		font-weight: 500;
-	}
-
-	.btn {
-		padding: 0.5rem 1rem;
-		font-size: 0.875rem;
-		font-weight: 500;
-		border: none;
-		border-radius: 6px;
-		cursor: pointer;
-		transition: all 0.2s;
-		white-space: nowrap;
-	}
-
-	.btn:disabled {
-		opacity: 0.5;
-		cursor: not-allowed;
-	}
-
-	.btn-primary {
-		background-color: #4a9eff;
-		color: #fff;
-	}
-
-	.btn-primary:hover:not(:disabled) {
-		background-color: #6bb3ff;
-	}
-
-	.btn-secondary {
-		background-color: #333;
-		color: #fff;
-	}
-
-	.btn-secondary:hover {
-		background-color: #444;
-	}
-
-	.btn-danger {
-		background-color: #ff4444;
-		color: #fff;
-	}
-
-	.btn-danger:hover {
-		background-color: #ff6666;
-	}
-
 	.viewer-wrapper {
 		flex: 1;
 		overflow: hidden;
@@ -485,111 +188,6 @@ $: if (!$collaborationStore.isInSession) {
 		justify-content: center;
 		padding: 2rem;
 		background-color: #1a1a1a;
-	}
-
-	.loading-overlay {
-		position: absolute;
-		inset: 0;
-		display: flex;
-		align-items: center;
-		justify-content: center;
-		background-color: rgba(26, 26, 26, 0.95);
-		z-index: 100;
-	}
-
-	.loading-content {
-		text-align: center;
-		max-width: 400px;
-	}
-
-	.spinner {
-		width: 48px;
-		height: 48px;
-		border: 4px solid #333;
-		border-top-color: #4a9eff;
-		border-radius: 50%;
-		animation: spin 1s linear infinite;
-		margin: 0 auto 1rem;
-	}
-
-	@keyframes spin {
-		to {
-			transform: rotate(360deg);
-		}
-	}
-
-	.loading-message {
-		margin: 0 0 1rem 0;
-		font-size: 1rem;
-		color: #ccc;
-	}
-
-	.progress-bar {
-		width: 100%;
-		height: 8px;
-		background-color: #333;
-		border-radius: 4px;
-		overflow: hidden;
-		margin-bottom: 0.5rem;
-	}
-
-	.progress-fill {
-		height: 100%;
-		background-color: #4a9eff;
-		transition: width 0.1s linear;
-	}
-
-	.progress-text {
-		margin: 0;
-		font-size: 0.875rem;
-		color: #888;
-	}
-
-	.error-overlay {
-		position: absolute;
-		top: 1rem;
-		right: 1rem;
-		z-index: 200;
-	}
-
-	.error-content {
-		background-color: #3a1a1a;
-		border: 1px solid #ff4444;
-		border-radius: 8px;
-		padding: 1rem 1.5rem;
-		max-width: 500px;
-		max-height: 80vh;
-		overflow-y: auto;
-	}
-
-	.error-title {
-		margin: 0 0 0.5rem 0;
-		font-size: 1rem;
-		font-weight: 600;
-		color: #ff6666;
-	}
-
-	.error-message {
-		margin: 0 0 1rem 0;
-		font-size: 0.875rem;
-		color: #ffaaaa;
-		white-space: pre-wrap;
-		line-height: 1.5;
-	}
-
-	.error-button {
-		background-color: #ff4444;
-		color: #fff;
-		border: none;
-		border-radius: 4px;
-		padding: 0.5rem 1rem;
-		font-size: 0.875rem;
-		cursor: pointer;
-		transition: background-color 0.2s;
-	}
-
-	.error-button:hover {
-		background-color: #ff6666;
 	}
 
 	.controls-info {
@@ -613,93 +211,10 @@ $: if (!$collaborationStore.isInSession) {
 		text-decoration: underline;
 	}
 
-	.qr-code-panel {
-		position: fixed;
-		top: 50%;
-		left: 50%;
-		transform: translate(-50%, -50%);
-		padding: 1.5rem;
-		background: rgba(0, 0, 0, 0.95);
-		border: 1px solid #444;
-		border-radius: 8px;
-		z-index: 1000;
-		box-shadow: 0 4px 20px rgba(0, 0, 0, 0.5);
-		backdrop-filter: blur(4px);
-		animation: fadeIn 0.2s ease-out;
-	}
-
-	@keyframes fadeIn {
-		from {
-			opacity: 0;
-			transform: translate(-50%, -50%) scale(0.95);
-		}
-		to {
-			opacity: 1;
-			transform: translate(-50%, -50%) scale(1);
-		}
-	}
-
-	.qr-code-content {
-		display: flex;
-		flex-direction: column;
-		align-items: center;
-		gap: 1rem;
-	}
-
-	.qr-code-title {
-		margin: 0;
-		font-size: 1rem;
-		font-weight: 600;
-		color: #fff;
-	}
-
-	.qr-code-image {
-		width: 300px;
-		height: 300px;
-		border: 4px solid #fff;
-		border-radius: 8px;
-		background-color: #fff;
-	}
-
-	.qr-code-hint {
-		margin: 0;
-		font-size: 0.875rem;
-		color: #888;
-		text-align: center;
-		max-width: 300px;
-	}
-
-	.qr-code-loading {
-		margin: 0;
-		font-size: 0.875rem;
-		color: #4a9eff;
-		padding: 2rem;
-	}
-
 	/* Hide keyboard shortcuts on mobile (use FAB instead), but keep footer note */
 	@media (max-width: 1023px) {
 		.keyboard-shortcuts {
 			display: none;
-		}
-
-		.header-content {
-			flex-direction: column;
-			align-items: flex-start;
-			gap: 1rem;
-		}
-
-		.session-controls {
-			width: 100%;
-			flex-wrap: wrap;
-		}
-
-		.session-info {
-			flex: 1;
-			min-width: 200px;
-		}
-
-		.btn {
-			flex: 1;
 		}
 	}
 </style>
